model/requirements.txt
PyQt5
numpy
pandas
torch
opencv-python
pillow
torchvision
matplotlib
seaborn
scikit-learn 